Fix addronly-reg test on Windows.
Fixes issue 649.

In this test, we need to perform an unaddressable memory access that
partially writes aflags.  When run natively, that messes up the
allocator's headers.  glibc malloc in particular doesn't like this, and
will abort on the call to free.  To avoid this, we were just leaking the
memory.

On Win7, however, in the addronly-reg test, the leak checker was finding
a stale pointer to this allocation, and not counting it as a leak.  My
theory is that in the process of reporting the unaddressable accesses,
the access base register is copied many times.  Some copy is left on the
stack when the application exits, and so the leak is not reported.  This
doesn't happen for other memory that we leak because they are allocated
earlier in the program and are overwritten.

Eventually, to avoid this kind of false negative, we should zero out
stack frames as is done in -leaks_only mode, which is covered by i#624.

To make the test pass both natively and have predictable output, we now
undo the modifications that we make to the headers and free the memory.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@574 5b54908e-ead4-fc18-b48a-0285e6dbc711