Make module name matching case insensitive with native Windows Perl.
In the suppress test on Cygwin, postprocess.pl was not uppercasing the
module names in the suppression files because it is using native Windows
Perl and the case insensitivity is guarded by $is_cygwin.  I've changed
the conditional to use !$is_unix instead, since this should cover both
native Windows apps and Cygwin apps.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@597 5b54908e-ead4-fc18-b48a-0285e6dbc711