Buffer writes to a temp symcache file and rename it into place.
Part of i#692: The unit_tests.exe symcache on the bot was getting larger
than 128K due to post-call addrs.  This removes the static buffer and
uses the buffered file writing utilities to print a file of arbitrary
size.

This also fixes some bugs in the BUFFERED_WRITE and FLUSH_BUFFER macros
that were introduced when changing dr_snprintf's return value.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@631 5b54908e-ead4-fc18-b48a-0285e6dbc711