i#707 and i#664: Accept DrMemory's output for cs2bug on Cygwin.
We now accept 4 invalid heap arguments, and 2 distinct leaks.

We don't have nested pattern matching, so I opted to use ANYLINE to
accept 3, 4, or 5 invalid heap args on any platform.  Similarly for the
extra leak we fail to suppress.  We'll let the .res file do the platform
dependent checking, and allow the extra system leak.

The %ANYLINE pattern messes up the code that looks for ~~Dr\.M~~ lines
to ignore, so I added a separate heapstat.out file to work around the
problem.

I also found a bug in runtest.cmake with %ANYLINE that has been
preventing the line after %ENDANYLINE from matching.  We were replacing
the trailing \n with |, so the next line became one of the match
alternatives.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@649 5b54908e-ead4-fc18-b48a-0285e6dbc711