Add support for VALGRIND_MAKE_MEM_DEFINED_IF_ADDRESSABLE.
This is the beginning of general support for Valgrind's client requests
mechanism.  At some point, we may want to lower this back into DynamoRIO so that
DynamoRIO can support the DISCARD_TRANSLATIONS annotation for better performance
on JITs.

For now, this should help us make the full DrMemory Chrome bot greener.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@676 5b54908e-ead4-fc18-b48a-0285e6dbc711