i#635 and i#603: Print offsets in stacks if there were no symbols.
This behavior is consistent with windbg and allows the user to tell if
we were able to find debug symbols for the module or not.  It also
provides valuable information for figuring out if the symbol that was
found (most likely an export) really contains the PC.

Fixes issue 635.
Fixes issue 603.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@684 5b54908e-ead4-fc18-b48a-0285e6dbc711