Add tests for and suppress leaks in ole32.dll from CoInitialize(Ex).
Fixes issue 18.

The first leak is described in comment 8.  EventPoolEntry::CreatePoolEntry
allocates a linked list node using SLIST_ENTRY, but the entry struct is in the
middle of the struct, so the head of the list points into the middle of the
chunk.  It may be possible to detect this kind of list if it has more than one
element, but with only one element it looks like all other allocations with a
mid-chunk pointer, so we suppress it.

The second leak is described in comment 9.  This looks like a true one time leak
in COleTls::TLSAllocData.

Unfortunately, until we have some better way to match suppressions when we don't
have pdbs (i#741), we have to use <mod+offs>.  This is likely to be brittle.  If
we see any reports on the bots or on test machines, we can keep adding the
offsets to the default suppression set.