i#337: Add an undefined exception for RtlpLowFragHeapAllocFromContext.
The exception checks if we're in a heap routine and if the current PC
and retaddr are both in ntdll.

We have a default suppression for this, but it means every time the app
does a low fragmentation malloc, which is every malloc in Chrome, we
have to at least take a callstack and match it with a recorded
callstack.  For every unique callstack, we will fully symbolize it
before suppressing it and saving the packed callstack.  Avoiding this
should hopefully save memory.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@848 5b54908e-ead4-fc18-b48a-0285e6dbc711