fixes issue 963
Adds -skip_msvc_importers option that checks whether a module (that is
not msvcp*.dll) imports from msvc*.dll.  If it does, the module is not
checked for malloc-related symbols or for string or memory related
symbols.  It is still checked for C++ operators and for
std::_DebugHeapDelete<*> as both are needed for detecting heap
mismatches.  This shrinks the unit_tests component load time from 2
minutes to 40 seconds in my tests.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@970 5b54908e-ead4-fc18-b48a-0285e6dbc711