i#945: 64-bit windows app system calls, part 1
Changed the tables to store the arg count rather than the arg size which
makes for a big diff but will be much less confusing long-term.

Also updated the script, and found and fixed some bugs in the table where
flags were added to the count instead of the flags field!

git-svn-id: https://drmemory.googlecode.com/svn/trunk@1020 5b54908e-ead4-fc18-b48a-0285e6dbc711