i#1014: Add a cpackappend= arg to package.cmake
This gives the caller of package.cmake an escape hatch to pass arbitrary
options to cpack.

My intention is to use it like this on the bots:
cpackappend='set(CPACK_PACKAGE_FILE_NAME "DrMemory-package")'

By setting the package name, I can make the directory that cpack uses to
stage the .zip archive predictable.  I can use that to make a 7-zip
self-extracting exe of it instead of a zipfile, which is what the Chrome
scripts expect.

Review URL: https://codereview.appspot.com/6499085

git-svn-id: https://drmemory.googlecode.com/svn/trunk@1021 5b54908e-ead4-fc18-b48a-0285e6dbc711