i#1057: Switch to dr_sscanf() in drmemory to avoid problems with ntdll
One issue is that dr_sscanf() does not support scan sets, which the
symcache was using to match symbols up to the next comma.  I converted
that call site to find the comma with strchr rather than implement scan
sets in dr_sscanf().

For consistency, I also converted our other internal uses of sscanf() to
dr_sscanf().  This lets us get rid of the special _GNU_SOURCE inclusions
of stdio.h that we have for issue 344.