i#1089: add inlined types for ntoskrnl pre-XP64 syscalls.
Bumped up and renamed the limit on stored args to reflect that we are no
longer storing just memory args.

Fixed bugs on arg iteration and on size entry location: the max entry is not
at arg_count index because of double entries.

Added 9 new headers in wininc/ to include defines for all of the new
inlined types (many are information class enums).

git-svn-id: https://drmemory.googlecode.com/svn/trunk@1180 5b54908e-ead4-fc18-b48a-0285e6dbc711