i#566: drstrace
Adds an initial version of an strace tool.  It uses drsyscall with its
recently added inlined arg information (i#1089) to print out system call
args and return values.

There are a lot of features that could be added.  Once a little more is in
place we can close this issue and open new ones for optional features.  The
rest of i#1089 is also needed for full arg info for all syscalls.

Currently the frontend is pulled out of drmemory/frontend.c.  I filed
DRi#1079 for sharing as much as possible.  I added workarounds for DRi#1082
to both frontends.

I added a sanity check test for drstrace.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@1187 5b54908e-ead4-fc18-b48a-0285e6dbc711