fixes issue 1148
Breaks drsyscall compatibility: but drsyscall has not been officially
released yet so that's allowed.  This commit message is a heads up to
anyone using it now.

Adds is_byte_undefined() to drsys_options_t, because undefined != !defined
and we want to avoid the perf hit of two shadow lookups.

Also relaxes assert that could fire in a race to a warning.

git-svn-id: https://drmemory.googlecode.com/svn/trunk@1209 5b54908e-ead4-fc18-b48a-0285e6dbc711